Rewrote termios constants because some platforms don't define them properly; basically, if I don't see a contast then I guess. Most of the constant values do seem to be the same across all UNIXes.
Got rid of lambda when build 'name' attribute in init.

Changed init so that 'command' attribute gets reset with the path and
so I can use execv instead of execvp. It seems that execvp was giving
weird results on some pythons and/or platforms. Since I now figure out
the path myself this should be more consistent across platforms.

Improved debugging output on test_expect.

Discovered surprising side-effect of using buffered input.
That is that now regexes in the list are matched first in the order
of the regex in the pattern list and not in order seen in the stream.
One would expect that the first regex in the stream would be matched first.
The new result makes sense if you understand how the scan works, but
it does not follow the priniple of least surprise. I don't like this.
I need to add some code to check to make sure that a match is also
the first in the stream.